requirements.txt
python-dotenv>=1.0.1
discord>=2.3.2
google-generativeai>=0.8.4